move prefilter code to pkg/datapath
This code was in `pkg/policy` but is used for configuring the datapath
directly, and does not tie into the policy repository. Refactor it into
`pkg/datapath` accordingly.

Fixes: #5865

Signed-off by: Ian Vernon <ian@cilium.io>
